requirements.txt
Flask
ElementTree
Jinja2
argh
BeautifulSoup
simplejson
isodate
pyserial
nose
